Refactor: inline paths in view, rename component and specs / change back link paths
- Inline path variables in accredited provider view templates
- Rename accredited_provider_details.rb and move it out of nested directory
- Rename spec file to accredited_provider_details_spec.rb
- Ensure backlinks on can_remove and cannot_remove page redirect to the accredited provider details page
