+++ app/components/publish/providers/accredited_provider_details.erb
