--- app/components/publish/providers/accredited_provider_details/view.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-module Publish
-  module Providers
-    module AccreditedProviderDetails
-      class View < ApplicationComponent
-        include Publish::ValueHelper
-
-        def initialize(provider:, classes: [], html_attributes: {})
-          super(classes:, html_attributes:)
-          @provider = provider
-        end
-
-        def formatted_provider_type
-          Provider.human_attribute_name(@provider.provider_type)
-        end
-
-        def formatted_accrediting_provider
-          @provider.accredited? ? "Yes" : "No"
-        end
-      end
-    end
-  end
-end
