+++ app/views/publish/providers/accredited_partnerships/_can_remove.html.erb
@@ -1,6 +1,12 @@
 <%= content_for :page_title, t(".title") %>
 <% content_for :before_content do %>
-  <%= govuk_back_link_to(publish_provider_recruitment_cycle_accredited_partnerships_path) %>
+  <%= render GovukComponent::BackLinkComponent.new(
+        href: publish_provider_recruitment_cycle_accredited_partnership_path(
+        provider_code: @provider_partnership.training_provider.provider_code,
+        recruitment_cycle_year: @provider_partnership.accredited_provider.recruitment_cycle_year,
+        accredited_provider_code: @provider_partnership.accredited_provider.provider_code,
+      ),
+      ) %>
 <% end %>
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-two-thirds">
