+++ app/views/publish/providers/accredited_partnerships/show.html.erb
@@ -1,26 +1,22 @@
-<% accredited_provider_list_path = publish_provider_recruitment_cycle_accredited_partnerships_path(
-     provider_code: @training_provider.provider_code,
-     recruitment_cycle_year: @training_provider.recruitment_cycle_year,
-   ) %>
-<% delete_accredited_provider_path = delete_publish_provider_recruitment_cycle_accredited_partnership_path(
-     accredited_provider_code: @accredited_provider.provider_code,
-     provider_code: @training_provider.provider_code,
-     recruitment_cycle_year: @training_provider.recruitment_cycle_year,
-   ) %>
-
 <% content_for :page_title, "#{@accredited_provider.provider_name} (#{@accredited_provider.provider_code})" %>
 <% content_for :before_content do %>
   <%= render GovukComponent::BackLinkComponent.new(
-        text: t(".back_link_text"),
-        href: accredited_provider_list_path,
+        href: publish_provider_recruitment_cycle_accredited_partnerships_path(
+          provider_code: @training_provider.provider_code,
+          recruitment_cycle_year: @training_provider.recruitment_cycle_year,
+        ),
       ) %>
 <% end %>
 <h1 class="govuk-heading-l">
   <%= @accredited_provider.provider_name %>
 </h1>
-<%= render Publish::Providers::AccreditedProviderDetails::View.new(provider: @accredited_provider) %>
+<%= render Publish::Providers::AccreditedProviderDetails.new(provider: @accredited_provider) %>
 <p class="govuk-body">
   <%= govuk_link_to t(".remove"),
-            delete_accredited_provider_path,
-            class: "app-link--destructive govuk-!-font-weight-regular" %>
+        delete_publish_provider_recruitment_cycle_accredited_partnership_path(
+          accredited_provider_code: @accredited_provider.provider_code,
+          provider_code: @training_provider.provider_code,
+          recruitment_cycle_year: @training_provider.recruitment_cycle_year,
+        ),
+        class: "app-link--destructive govuk-!-font-weight-regular" %>
 </p>
